requirements.txt
albumentations==1.3.0
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docker-pycreds==0.4.0
filelock==3.12.2
gitdb==4.0.10
GitPython==3.1.31
h5py==3.9.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.0
sentry-sdk==1.26.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==1.12.1
torchstat==0.0.7
torchvision==0.13.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wandb==0.15.2
